Return telemetry and some other things into the common azure plugin
