+++ PluginsAndFeatures/azure-toolkit-for-intellij/src/com/microsoft/intellij/AzurePlugin.java
@@ -34,35 +34,52 @@ import com.intellij.util.PlatformUtils;
 import com.intellij.util.containers.HashSet;
 import com.microsoft.applicationinsights.preference.ApplicationInsightsResource;
 import com.microsoft.applicationinsights.preference.ApplicationInsightsResourceRegistry;
+import com.microsoft.azuretools.authmanage.CommonSettings;
 import com.microsoft.azuretools.azurecommons.deploy.DeploymentEventArgs;
 import com.microsoft.azuretools.azurecommons.deploy.DeploymentEventListener;
+import com.microsoft.azuretools.azurecommons.helpers.StringHelper;
 import com.microsoft.azuretools.azurecommons.util.FileUtil;
+import com.microsoft.azuretools.azurecommons.util.GetHashMac;
+import com.microsoft.azuretools.azurecommons.util.ParserXMLUtility;
 import com.microsoft.azuretools.azurecommons.util.Utils;
 import com.microsoft.azuretools.azurecommons.util.WAEclipseHelperMethods;
+import com.microsoft.azuretools.azurecommons.xmlhandling.DataOperations;
+import com.microsoft.azuretools.telemetry.AppInsightsClient;
+import com.microsoft.azuretools.telemetry.AppInsightsConstants;
+import com.microsoft.azuretools.utils.TelemetryUtils;
 import com.microsoft.intellij.common.CommonConst;
 import com.microsoft.intellij.ui.libraries.AILibraryHandler;
 import com.microsoft.intellij.ui.libraries.AzureLibrary;
 import com.microsoft.intellij.ui.messages.AzureBundle;
+import com.microsoft.intellij.util.PluginHelper;
 import com.microsoft.intellij.util.PluginUtil;
+import org.apache.commons.io.FileUtils;
+import org.w3c.dom.Document;
 
 import javax.swing.event.EventListenerList;
+import java.io.BufferedOutputStream;
 import java.io.File;
+import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
+import java.net.URL;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
+import java.util.zip.ZipEntry;
+import java.util.zip.ZipInputStream;
 
 import static com.microsoft.intellij.ui.messages.AzureBundle.message;
 
 
-public abstract class AzurePlugin extends AbstractProjectComponent {
+public class AzurePlugin extends AbstractProjectComponent {
     protected static final Logger LOG = Logger.getInstance("#com.microsoft.intellij.AzurePlugin");
 
     public static final String PLUGIN_VERSION = CommonConst.PLUGIN_VERISON;
     public static final String AZURE_LIBRARIES_VERSION = "1.0.0";
     public static final String JDBC_LIBRARIES_VERSION = "6.1.0.jre8";
+    public static final int REST_SERVICE_MAX_RETRY_COUNT = 7;
 
     // User-agent header for Azure SDK calls
     public static final String USER_AGENT = "Azure Toolkit for IntelliJ, v%s, machineid:%s";
@@ -72,15 +89,22 @@ public abstract class AzurePlugin extends AbstractProjectComponent {
 
     public static String pluginFolder = PluginUtil.getPluginRootDirectory();
 
-    protected static final EventListenerList DEPLOYMENT_EVENT_LISTENERS = new EventListenerList();
+    private static final EventListenerList DEPLOYMENT_EVENT_LISTENERS = new EventListenerList();
     public static List<DeploymentEventListener> depEveList = new ArrayList<DeploymentEventListener>();
 
+    protected final String dataFile = PluginHelper.getTemplateFile(message("dataFileName"));
+
     private final AzureSettings azureSettings;
+
+    private String _hashmac = GetHashMac.GetHashMac();
 
     public AzurePlugin(Project project) {
         super(project);
         this.azureSettings = AzureSettings.getSafeInstance(project);
+        CommonSettings.setUserAgent(String.format(USER_AGENT, PLUGIN_VERSION,
+                TelemetryUtils.getMachieId(dataFile, message("prefVal"), message("instID"))));
     }
+
 
     public void projectOpened() {
         initializeAIRegistry();
@@ -110,7 +134,52 @@ public abstract class AzurePlugin extends AbstractProjectComponent {
         }
     }
 
-    protected void initializeTelemetry() throws Exception {
+    private void initializeTelemetry() throws Exception {
+        boolean install = false;
+        boolean upgrade = false;
+
+        if (new File(dataFile).exists()) {
+            String version = DataOperations.getProperty(dataFile, message("pluginVersion"));
+            if (version == null || version.isEmpty()) {
+                upgrade = true;
+                // proceed with setValues method as no version specified
+                setValues(dataFile);
+            } else {
+                String curVersion = PLUGIN_VERSION;
+                // compare version
+                if (curVersion.equalsIgnoreCase(version)) {
+                    // Case of normal IntelliJ restart
+                    // check preference-value & installation-id exists or not else copy values
+                    String prefValue = DataOperations.getProperty(dataFile, message("prefVal"));
+                    String instID = DataOperations.getProperty(dataFile, message("instID"));
+                    if (prefValue == null || prefValue.isEmpty()) {
+                        setValues(dataFile);
+                    } else if (instID == null || instID.isEmpty() || !GetHashMac.IsValidHashMacFormat(instID)) {
+                        upgrade = true;
+                        Document doc = ParserXMLUtility.parseXMLFile(dataFile);
+                        DataOperations.updatePropertyValue(doc, message("instID"), _hashmac);
+                        ParserXMLUtility.saveXMLFile(dataFile, doc);
+                    }
+                } else {
+                    upgrade = true;
+                    // proceed with setValues method. Case of new plugin installation
+                    setValues(dataFile);
+                }
+            }
+        } else {
+            // copy file and proceed with setValues method
+            install = true;
+            copyResourceFile(message("dataFileName"), dataFile);
+            setValues(dataFile);
+        }
+        AppInsightsClient.setAppInsightsConfiguration(new AppInsightsConfigurationImpl());
+        if (install) {
+            AppInsightsClient.createByType(AppInsightsClient.EventType.Plugin, "", AppInsightsConstants.Install, null, true);
+        }
+        if (upgrade) {
+            AppInsightsClient.createByType(AppInsightsClient.EventType.Plugin, "", AppInsightsConstants.Upgrade, null, true);
+        }
+        AppInsightsClient.createByType(AppInsightsClient.EventType.Plugin, "", AppInsightsConstants.Load, null, true);
     }
 
     private void initializeAIRegistry() {
@@ -140,6 +209,23 @@ public abstract class AzurePlugin extends AbstractProjectComponent {
             AzureSettings.getSafeInstance(myProject).saveAppInsights();
         } catch (Exception ex) {
             AzurePlugin.log(ex.getMessage(), ex);
+        }
+    }
+
+    private void setValues(final String dataFile) throws Exception {
+        try {
+            final Document doc = ParserXMLUtility.parseXMLFile(dataFile);
+            String recordedVersion = DataOperations.getProperty(dataFile, message("pluginVersion"));
+            if (Utils.whetherUpdateTelemetryPref(recordedVersion)) {
+                DataOperations.updatePropertyValue(doc, message("prefVal"), String.valueOf("true"));
+            }
+
+            DataOperations.updatePropertyValue(doc, message("pluginVersion"), PLUGIN_VERSION);
+            DataOperations.updatePropertyValue(doc, message("instID"), _hashmac);
+
+            ParserXMLUtility.saveXMLFile(dataFile, doc);
+        } catch (Exception ex) {
+            LOG.error(message("error"), ex);
         }
     }
 
@@ -185,8 +271,18 @@ public abstract class AzurePlugin extends AbstractProjectComponent {
                 });
     }
 
+    private void telemetryAI() {
+        ModuleManager.getInstance(myProject).getModules();
+    }
+
     public String getComponentName() {
         return "MSOpenTechTools.AzurePlugin";
+    }
+
+    // currently we didn't have a better way to know if it is in debug model.
+    // the code suppose we are under debug model if the plugin root path contains 'sandbox' for Gradle default debug path
+    protected boolean isDebugModel() {
+        return PluginUtil.getPluginRootDirectory().contains("sandbox");
     }
 
     /**
